extract custom minusInfinityDb into shared config

+++ Source/DecibelConfig.h
@@ -1,6 +1,6 @@
 #pragma once
 
 namespace DecibelConfig {
-    const double minusInfinityDb = -96;
+    const float minusInfinityDb = -96;
     const String customMinusInfinityString = "-INF";
 }
